+++ frontend/.expo/types/router.d.ts
@@ -6,9 +6,9 @@ export * from 'expo-router';
 declare module 'expo-router' {
   export namespace ExpoRouter {
     export interface __routes<T extends string | object = string> {
-      hrefInputParams: { pathname: Router.RelativePathString, params?: Router.UnknownInputParams } | { pathname: Router.ExternalPathString, params?: Router.UnknownInputParams } | { pathname: `/_sitemap`; params?: Router.UnknownInputParams; } | { pathname: `${'/(auth)'}/login` | `/login`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/archived` | `/archived`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/taobaiviet` | `/taobaiviet`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/todo` | `/todo`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}/homeScreen` | `/homeScreen`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}` | `/`; params?: Router.UnknownInputParams; } | { pathname: `/+not-found`, params: Router.UnknownInputParams & {  } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]` | `/lopHocPhan/[id]`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/dashboard` | `/lopHocPhan/[id]/dashboard`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/notifications` | `/lopHocPhan/[id]/notifications`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/peopleScreen` | `/lopHocPhan/[id]/peopleScreen`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/baiviet/[id]` | `/lopHocPhan/[id]/baiviet/[id]`, params: Router.UnknownInputParams & { id: string | number;id: string | number; } };
-      hrefOutputParams: { pathname: Router.RelativePathString, params?: Router.UnknownOutputParams } | { pathname: Router.ExternalPathString, params?: Router.UnknownOutputParams } | { pathname: `/_sitemap`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(auth)'}/login` | `/login`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}/archived` | `/archived`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}/taobaiviet` | `/taobaiviet`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}/todo` | `/todo`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}/homeScreen` | `/homeScreen`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}` | `/`; params?: Router.UnknownOutputParams; } | { pathname: `/+not-found`, params: Router.UnknownOutputParams & {  } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]` | `/lopHocPhan/[id]`, params: Router.UnknownOutputParams & { id: string; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/dashboard` | `/lopHocPhan/[id]/dashboard`, params: Router.UnknownOutputParams & { id: string; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/notifications` | `/lopHocPhan/[id]/notifications`, params: Router.UnknownOutputParams & { id: string; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/peopleScreen` | `/lopHocPhan/[id]/peopleScreen`, params: Router.UnknownOutputParams & { id: string; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/baiviet/[id]` | `/lopHocPhan/[id]/baiviet/[id]`, params: Router.UnknownOutputParams & { id: string;id: string; } };
-      href: Router.RelativePathString | Router.ExternalPathString | `/_sitemap${`?${string}` | `#${string}` | ''}` | `${'/(auth)'}/login${`?${string}` | `#${string}` | ''}` | `/login${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}/archived${`?${string}` | `#${string}` | ''}` | `/archived${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}/taobaiviet${`?${string}` | `#${string}` | ''}` | `/taobaiviet${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}/todo${`?${string}` | `#${string}` | ''}` | `/todo${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}${'/(tabs)'}/homeScreen${`?${string}` | `#${string}` | ''}` | `/homeScreen${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}${'/(tabs)'}${`?${string}` | `#${string}` | ''}` | `/${`?${string}` | `#${string}` | ''}` | { pathname: Router.RelativePathString, params?: Router.UnknownInputParams } | { pathname: Router.ExternalPathString, params?: Router.UnknownInputParams } | { pathname: `/_sitemap`; params?: Router.UnknownInputParams; } | { pathname: `${'/(auth)'}/login` | `/login`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/archived` | `/archived`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/taobaiviet` | `/taobaiviet`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/todo` | `/todo`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}/homeScreen` | `/homeScreen`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}` | `/`; params?: Router.UnknownInputParams; } | `/+not-found` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}` | `/lopHocPhan/${Router.SingleRoutePart<T>}` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}${'/(tabs)'}/dashboard` | `/lopHocPhan/${Router.SingleRoutePart<T>}/dashboard` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}${'/(tabs)'}/notifications` | `/lopHocPhan/${Router.SingleRoutePart<T>}/notifications` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}${'/(tabs)'}/peopleScreen` | `/lopHocPhan/${Router.SingleRoutePart<T>}/peopleScreen` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}${'/(tabs)'}/baiviet/${Router.SingleRoutePart<T>}` | `/lopHocPhan/${Router.SingleRoutePart<T>}/baiviet/${Router.SingleRoutePart<T>}` | { pathname: `/+not-found`, params: Router.UnknownInputParams & {  } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]` | `/lopHocPhan/[id]`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/dashboard` | `/lopHocPhan/[id]/dashboard`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/notifications` | `/lopHocPhan/[id]/notifications`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/peopleScreen` | `/lopHocPhan/[id]/peopleScreen`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]${'/(tabs)'}/baiviet/[id]` | `/lopHocPhan/[id]/baiviet/[id]`, params: Router.UnknownInputParams & { id: string | number;id: string | number; } };
+      hrefInputParams: { pathname: Router.RelativePathString, params?: Router.UnknownInputParams } | { pathname: Router.ExternalPathString, params?: Router.UnknownInputParams } | { pathname: `/_sitemap`; params?: Router.UnknownInputParams; } | { pathname: `${'/(auth)'}/login` | `/login`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/archived` | `/archived`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/taobaiviet` | `/taobaiviet`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/todo` | `/todo`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}/homeScreen` | `/homeScreen`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}` | `/`; params?: Router.UnknownInputParams; } | { pathname: `/+not-found`, params: Router.UnknownInputParams & {  } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/baiviet/[id]` | `/lopHocPhan/[id]/baiviet/[id]`, params: Router.UnknownInputParams & { id: string | number;id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/classTabScreen` | `/lopHocPhan/[id]/class(tabs)/classTabScreen`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/dashboard` | `/lopHocPhan/[id]/class(tabs)/dashboard`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/notifications` | `/lopHocPhan/[id]/class(tabs)/notifications`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/peopleScreen` | `/lopHocPhan/[id]/class(tabs)/peopleScreen`, params: Router.UnknownInputParams & { id: string | number; } };
+      hrefOutputParams: { pathname: Router.RelativePathString, params?: Router.UnknownOutputParams } | { pathname: Router.ExternalPathString, params?: Router.UnknownOutputParams } | { pathname: `/_sitemap`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(auth)'}/login` | `/login`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}/archived` | `/archived`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}/taobaiviet` | `/taobaiviet`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}/todo` | `/todo`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}/homeScreen` | `/homeScreen`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}` | `/`; params?: Router.UnknownOutputParams; } | { pathname: `/+not-found`, params: Router.UnknownOutputParams & {  } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/baiviet/[id]` | `/lopHocPhan/[id]/baiviet/[id]`, params: Router.UnknownOutputParams & { id: string;id: string; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/classTabScreen` | `/lopHocPhan/[id]/class(tabs)/classTabScreen`, params: Router.UnknownOutputParams & { id: string; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/dashboard` | `/lopHocPhan/[id]/class(tabs)/dashboard`, params: Router.UnknownOutputParams & { id: string; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/notifications` | `/lopHocPhan/[id]/class(tabs)/notifications`, params: Router.UnknownOutputParams & { id: string; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/peopleScreen` | `/lopHocPhan/[id]/class(tabs)/peopleScreen`, params: Router.UnknownOutputParams & { id: string; } };
+      href: Router.RelativePathString | Router.ExternalPathString | `/_sitemap${`?${string}` | `#${string}` | ''}` | `${'/(auth)'}/login${`?${string}` | `#${string}` | ''}` | `/login${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}/archived${`?${string}` | `#${string}` | ''}` | `/archived${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}/taobaiviet${`?${string}` | `#${string}` | ''}` | `/taobaiviet${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}/todo${`?${string}` | `#${string}` | ''}` | `/todo${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}${'/(tabs)'}/homeScreen${`?${string}` | `#${string}` | ''}` | `/homeScreen${`?${string}` | `#${string}` | ''}` | `${'/(drawer)'}${'/(tabs)'}${`?${string}` | `#${string}` | ''}` | `/${`?${string}` | `#${string}` | ''}` | { pathname: Router.RelativePathString, params?: Router.UnknownInputParams } | { pathname: Router.ExternalPathString, params?: Router.UnknownInputParams } | { pathname: `/_sitemap`; params?: Router.UnknownInputParams; } | { pathname: `${'/(auth)'}/login` | `/login`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/archived` | `/archived`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/taobaiviet` | `/taobaiviet`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}/todo` | `/todo`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}/homeScreen` | `/homeScreen`; params?: Router.UnknownInputParams; } | { pathname: `${'/(drawer)'}${'/(tabs)'}` | `/`; params?: Router.UnknownInputParams; } | `/+not-found` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}/baiviet/${Router.SingleRoutePart<T>}` | `/lopHocPhan/${Router.SingleRoutePart<T>}/baiviet/${Router.SingleRoutePart<T>}` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}/class(tabs)/classTabScreen` | `/lopHocPhan/${Router.SingleRoutePart<T>}/class(tabs)/classTabScreen` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}/class(tabs)/dashboard` | `/lopHocPhan/${Router.SingleRoutePart<T>}/class(tabs)/dashboard` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}/class(tabs)/notifications` | `/lopHocPhan/${Router.SingleRoutePart<T>}/class(tabs)/notifications` | `${'/(drawer)'}${'/(class)'}/lopHocPhan/${Router.SingleRoutePart<T>}/class(tabs)/peopleScreen` | `/lopHocPhan/${Router.SingleRoutePart<T>}/class(tabs)/peopleScreen` | { pathname: `/+not-found`, params: Router.UnknownInputParams & {  } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/baiviet/[id]` | `/lopHocPhan/[id]/baiviet/[id]`, params: Router.UnknownInputParams & { id: string | number;id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/classTabScreen` | `/lopHocPhan/[id]/class(tabs)/classTabScreen`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/dashboard` | `/lopHocPhan/[id]/class(tabs)/dashboard`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/notifications` | `/lopHocPhan/[id]/class(tabs)/notifications`, params: Router.UnknownInputParams & { id: string | number; } } | { pathname: `${'/(drawer)'}${'/(class)'}/lopHocPhan/[id]/class(tabs)/peopleScreen` | `/lopHocPhan/[id]/class(tabs)/peopleScreen`, params: Router.UnknownInputParams & { id: string | number; } };
     }
   }
 }
