+++ frontend/app/(drawer)/(class)/lopHocPhan/[id]/_layout.tsx
@@ -1,20 +1,13 @@
-import { LopHocPhanProvider } from "@/context/_context";
 import { Stack } from "expo-router";
-import TabsLayout from "."; // 👈 import component Tabs
 
 export default function Layout() {
-  return (
-    <LopHocPhanProvider>
-      <Stack>
-        <Stack.Screen
-          name="(tabs)/dashboard"
-          options={{ title: "Bảng tin lớp học" }}
-        />
-        <Stack.Screen
-          name="(tabs)/baiviet/[id]"
-          options={{ title: "Chi tiết bài viết" }}
-        />
-      </Stack>
-    </LopHocPhanProvider>
-  );
+    return (
+        <Stack>
+            <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+            <Stack.Screen
+                name="baiviet/[id]"
+                options={{ title: "Chi tiết bài viết" }}
+            />
+        </Stack>
+    );
 }
