+++ frontend/app/(drawer)/(class)/lopHocPhan/[id]/class(tabs)/classTabScreen.tsx
@@ -1,0 +1,9 @@
+import { Redirect, useLocalSearchParams } from "expo-router";
+
+export default function ClassTabScreen() {
+    const { id } = useLocalSearchParams<{ id: string }>();
+
+    if (!id) return null;
+
+    return <Redirect href={`../classTabs/dashboard`} />;
+}
