+++ frontend/constants/api.ts
@@ -1,1 +1,1 @@
-export const API_BASE_URL = "http://192.168.1.102:3001";
+export const API_BASE_URL = "http://192.168.1.105:3001";
